+++ shared/versions.ts
@@ -1,5 +1,5 @@
-export const APP_VERSION = "2.3.98";
+export const APP_VERSION = "2.3.99";
 export const ANDROID_APP_VERSION = "2.3.34";
-export const ANDROID_APK_RELEASE_VERSION = "2.3.98";
+export const ANDROID_APK_RELEASE_VERSION = "2.3.99";
 export const AGENT_VERSION = "2.2.82";